src/requirements.txt
asgiref==3.7.1
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
django-guest-user==0.5.4
sqlparse==0.4.4
tzdata==2023.3
